refactor: update float32 decoder
